+++ 3.PresentationLayers/Views/Form_ChiTietTheLoai.resx
