+++ 3.PresentationLayers/Views/Form_NXB.resx
