+++ 3.PresentationLayers/Views/Form_TheLoai.resx
